Configures ddev for new project name "cms"
Updates ddev configuration to reflect the project rename from "webcms" to "cms".
This includes updating the project name in the ddev config file, as well as creating a new traefik configuration file for the new project.

Adds traefik configuration for routing HTTP and HTTPS traffic to the web and xhgui services.
This configuration includes routers, middlewares, and services definitions for both HTTP and HTTPS.